+++ NEWS.md
@@ -11,7 +11,23 @@
    `FutureJournalList` list of them now reports on parallelization
    metrics, including total speedup and total efficiency.
 
+ * `all.equal()` for `future` now returns `"Not both functions"` when
+   comparing against a non-function. Similarly, `"Not both lists"` is
+   returned when comparing a `FutureStrategyList` object against a not
+   list object.
+   
 ## Bug Fixes
+
+ * On MS Windows, a future startup script specified by environment
+   variable `R_FUTURE_STARTUP_SCRIPT` was silently ignored due to a
+   bug in parsing the value.
+
+ * Conditions relayed via the file system could produce an error when
+   relayed if a condition file was corrupt. Now such conditions are
+   silently ignored as intended.
+ 
+ * `all.equal()` for `FutureStrategyList` would throw an error if the
+   object checked against was not a list.
 
  * The default for R option `future.globals.maxSize` is now `+Inf`
    everywhere. It was still 500 MiB is some cases, e.g. `futureCall()`